+++ source/extensions/omni.flux.stage_manager.core/omni/flux/stage_manager/core/__init__.py
@@ -15,7 +15,7 @@
 * limitations under the License.
 """
 
-__all__ = ["StageManagerCore", "StageManagerSchema"]
+__all__ = ["StageManagerCore", "StageManagerSchema", "get_instance", "set_instance"]
 
-from .core import StageManagerCore
+from .core import StageManagerCore, get_instance, set_instance
 from .schema import StageManagerSchema