+++ source/extensions/omni.flux.stage_manager.plugin.interaction.usd/omni/flux/stage_manager/plugin/interaction/usd/all_tags.py
@@ -33,6 +33,7 @@ class AllTagsInteractionPlugin(_StageManagerUSDInteractionPlugin):
     )
     compatible_filters: list[str] = Field(
         default=[
+            "AdditionalFilterPlugin",
             "IgnorePrimsFilterPlugin",
             "IsCaptureFilterPlugin",
             "OmniPrimsFilterPlugin",