+++ source/extensions/lightspeed.trex.stage_manager.plugin.interaction.usd/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [2.2.0]
+### Added
+- Added `AdditionalFilterPlugin`
+
 ## [2.1.0]
 ### Added
 - Added `ParticleSystemsActionWidgetPlugin`