+++ source/extensions/omni.flux.stage_manager.factory/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [4.4.0]
+### Added
+- Added `additional_filters` to the interaction plugins
+
 ## [4.3.1]
 ### Fixed
 - Fixed stage manager not scrolling to the previous scroll position correctly